requirements.txt
click
git+https://github.com/feder-cr/lib_resume_builder_AIHawk.git
httpx
inputimeout
jsonschema
jsonschema-specifications
langchain
langchain-anthropic
langchain-huggingface
langchain-community
langchain-core
langchain-google-genai
langchain-ollama
langchain-openai
langchain-text-splitters
langsmith
Levenshtein
loguru
openai
pdfminer.six
pytest>=8.3.3
python-dotenv~=1.0.1
PyYAML~=6.0.2
regex
reportlab
selenium
webdriver-manager
pytest
pytest-mock
pytest-cov
undetected-chromedriver
google-generativeai==0.7.2